add NetworkInstance, Model and refactor some classes
