+++ app/src/main/java/kz/chocofamily/coroutinelesson/data/sources/NetworkInstance.kt
@@ -1,0 +1,19 @@
+package kz.chocofamily.coroutinelesson.data.sources
+
+import kz.chocofamily.coroutinelesson.data.configs.BASE_URL
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+
+object NetworkInstance {
+
+    private val retrofit: Retrofit = Retrofit.Builder()
+        .baseUrl(BASE_URL)
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+
+    val currencyApi: CurrencyApi by lazy {
+        retrofit.create(
+            CurrencyApi::class.java)
+    }
+}
