--- app/src/main/java/kz/chocofamily/coroutinelesson/sources/NetworkInstance.kt
@@ -1,10 +1,0 @@
-package kz.chocofamily.coroutinelesson.sources
-
-import retrofit2.Retrofit
-
-object NetworkInstance {
-
-    private val retrofit: Retrofit = Retrofit.Builder()
-        .addConverterFactory(Gson)
-        .build()
-}
